requirements.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.6.1.0
boto3==1.9.199
botocore==1.12.200
celery==4.3.0
certifi==2019.6.16
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.3
django-allauth==0.40.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-filters==0.2.1
django-leaflet==0.24.0
django-rest-auth==0.9.5
django-storages==1.7.1
django-widget-tweaks==1.4.5
djangorestframework==3.10.1
djangorestframework-gis==0.14
docutils==0.14
gunicorn==19.9.0
idna==2.8
importlib-metadata==0.23
jmespath==0.9.4
kombu==4.6.5
more-itertools==7.2.0
oauthlib==3.1.0
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2019.1
redis==3.3.6
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
vine==1.3.0
zipp==0.6.0
